Minor refactor of "Now listening" logs
Logs the actual address of the listener (rather than the configured string) so functionality tests can listen on port 0 and learn from the log which port the kernel chose.
Also reduces code duplication by logging at the start of the listener func instead of just before every call into it.